+++ tests/test_common.py
@@ -6,24 +6,24 @@ from compages._common import GeneratorStack, Result, get_lookup_order
 
 
 def test_normal_operation():
-    args = (1, 2)
-    stack = GeneratorStack(args, ["a"])
+    ref_context = "some context"
+    stack = GeneratorStack(ref_context, ["a"])
 
     assert stack.is_empty()
 
     # `None` is just ignored, and stack is not finalized
     assert stack.push(None) is Result.UNDEFINED
 
-    def f1(arg1, arg2, val):
-        assert (arg1, arg2) == args
+    def f1(context, val):
+        assert context == ref_context
         new_val = yield [*val, "b"]
         return [*new_val, "c"]
 
     # We pushed a generator, so the stack is not finalized yet
     assert stack.push(f1) is Result.UNDEFINED
 
-    def f2(arg1, arg2, val):
-        assert (arg1, arg2) == args
+    def f2(context, val):
+        assert context == ref_context
         return [*val, "d"]
 
     # This is a normal function, so the stack is finalized and unrolled
@@ -31,11 +31,11 @@ def test_normal_operation():
 
 
 def test_multiple_yields():
-    args = (1, 2)
-    stack = GeneratorStack(args, ["a"])
+    ref_context = "some context"
+    stack = GeneratorStack(ref_context, ["a"])
 
-    def f1(arg1, arg2, val):
-        assert (arg1, arg2) == args
+    def f1(context, val):
+        assert context == ref_context
         new_val = yield [*val, "b"]
         new_val = yield [*new_val, "d"]
         return [*val, "c"]
@@ -43,8 +43,8 @@ def test_multiple_yields():
     # We can't tell that there is a second yield until we start unrolling, so this passes
     stack.push(f1)
 
-    def f2(arg1, arg2, val):
-        assert (arg1, arg2) == args
+    def f2(context, val):
+        assert context == ref_context
         return [*val, "d"]
 
     with pytest.raises(RuntimeError, match="Expected only one yield in a generator"):